+++ app/decorators/facets/item_type_with_status.rb
@@ -1,7 +1,7 @@
 class Facets::ItemTypeWithStatus < Facets::DefaultFacetDecorator
 
   def display
-    I18n.t("controlled_vocabularies.item_type_with_status.#{@value}")
+    I18n.t("controlled_vocabularies.era.item_type_with_status.#{@value}")
   end
 
 end