+++ app/decorators/facets/visibility.rb
@@ -1,7 +1,7 @@
 class Facets::Visibility < Facets::URIValue
 
   def display
-    translate_uri(:visibility, @value)
+    translate_uri(:era, :visibility, @value)
   end
 
 end